Move initializer methods
